+++ volmdlr/edges.py
@@ -200,7 +200,7 @@ class Edge(dc.DessiaObject):
             if not same_sense:
                 obj = obj.reverse()
             if not point1.is_close(point2):
-                return LineSegment3D(point1, point2, obj, name=arguments[0][1:-1])
+                return LineSegment3D(point1, point2, obj, arguments[0][1:-1])
             return None
 
         if hasattr(obj, 'trim'):
@@ -996,7 +996,7 @@ class BSplineCurve(Edge):
         func, func_first_derivative, curve_derivatives, distance_vector = self._point_inversion_funcs(u0, point)
         if self._check_convergence(curve_derivatives, distance_vector, tol1=tol1, tol2=tol2):
             return u0, True
-        new_u = u0 - func / func_first_derivative
+        new_u = u0 - func / (func_first_derivative + 1e-18)
         new_u = self._check_bounds(new_u)
         residual = (new_u - u0) * curve_derivatives[1]
         if residual.norm() <= 1e-6:
@@ -1401,9 +1401,9 @@ class BSplineCurve(Edge):
         if all(other_bspline2.point_belongs(point, abs_tol=abs_tol) for point in self.points):
             return [self]
         if self.point_belongs(other_bspline2.start, abs_tol=abs_tol):
-            bspline1_, bspline2_ = self.split(other_bspline2.start)
+            bspline1_, bspline2_ = self.split(other_bspline2.start, tol=abs_tol)
         elif self.point_belongs(other_bspline2.end, abs_tol=abs_tol):
-            bspline1_, bspline2_ = self.split(other_bspline2.end)
+            bspline1_, bspline2_ = self.split(other_bspline2.end, tol=abs_tol)
         else:
             return []
             # raise NotImplementedError
@@ -1526,7 +1526,7 @@ class BSplineCurve(Edge):
         abscissa1 = self.abscissa(point1)
         abscissa2 = self.abscissa(point2)
         # special case periodical bsplinecurve
-        if self.periodic and abscissa2 == 0.0:
+        if self.periodic and math.isclose(abscissa2, 0.0, abs_tol=1e-6):
             abscissa2 = self.length()
         discretized_points_between_1_2 = []
         for abscissa in npy.linspace(abscissa1, abscissa2, num=number_points):
@@ -1553,6 +1553,15 @@ class BSplineCurve(Edge):
                 if is_true:
                     return True
         return False
+
+    def sort_points_along_curve(self, points: List[Union[volmdlr.Point2D, volmdlr.Point3D]]):
+        """
+        Sort point along a curve.
+
+        :param points: list of points to be sorted.
+        :return: sorted points.
+        """
+        return sorted(points, key=self.abscissa)
 
 
 class BSplineCurve2D(BSplineCurve):
@@ -2246,7 +2255,7 @@ class ArcMixin:
             if not angle_resolution:
                 number_points = 2
             else:
-                number_points = math.ceil(self.angle * angle_resolution) + 1
+                number_points = max(math.ceil(self.angle * angle_resolution) + 1, 2)
 
         step = self.length() / (number_points - 1)
         return [self.point_at_abscissa(i * step)
@@ -2386,7 +2395,7 @@ class FullArcMixin(ArcMixin):
     """
 
     def __init__(self, circle: Union[volmdlr.curves.Circle2D, volmdlr.curves.Circle3D],
-                 start_end: Union[volmdlr.Point2D, volmdlr.Point3D], name: str = ''):
+                 start_end: Union[volmdlr.Point2D, volmdlr.Point3D]):
         self.circle = circle
         self.start_end = start_end
         ArcMixin.__init__(self, circle=circle, start=start_end, end=start_end)  # !!! this is dangerous
@@ -2945,8 +2954,8 @@ class FullArc2D(FullArcMixin, Arc2D):
                  name: str = ''):
         # self.interior = start_end.rotation(center, math.pi)
         self._bounding_rectangle = None
-        FullArcMixin.__init__(self, circle=circle, start_end=start_end, name=name)
-        Arc2D.__init__(self, circle=circle, start=start_end, end=start_end)
+        FullArcMixin.__init__(self, circle=circle, start_end=start_end)
+        Arc2D.__init__(self, circle=circle, start=start_end, end=start_end, name=name)
         self.angle1 = 0.0
         self.angle2 = volmdlr.TWO_PI
 
@@ -3291,9 +3300,9 @@ class ArcEllipse2D(Edge):
                 abscissa_angle = initial_angle
                 break
             if res > abscissa:
-                increment_factor = (abs(initial_angle - angle_start) * (abscissa - res))/(2 * res)
+                increment_factor = (abs(initial_angle - angle_start) * (abscissa - res))/(2 * abs(res))
             else:
-                increment_factor = (abs(initial_angle - angle_start) * (abscissa - res))/res
+                increment_factor = (abs(initial_angle - angle_start) * (abscissa - res))/abs(res)
             initial_angle += increment_factor
             iter_counter += 1
         x = self.ellipse.major_axis * math.cos(abscissa_angle)
@@ -4021,6 +4030,9 @@ class LineSegment3D(LineSegment):
         return bspline_curve
 
     def get_reverse(self):
+        """
+        Gets the reverse of the Line Segment.
+        """
         return LineSegment3D(self.end.copy(), self.start.copy())
 
     def minimum_distance_points(self, other_line):
@@ -4053,7 +4065,7 @@ class LineSegment3D(LineSegment):
         return volmdlr.core_compiled.LineSegment3DDistance([self.start, self.end], [other_line.start, other_line.end])
 
     def parallel_distance(self, other_linesegment):
-        """Calculates the paralell distance between two Line Segments 3D."""
+        """Calculates the parallel distance between two Line Segments 3D."""
         pt_a, pt_b, pt_c = self.start, self.end, other_linesegment.start
         vector = volmdlr.Vector3D((pt_a - pt_b).vector)
         vector.normalize()
@@ -4346,15 +4358,13 @@ class BSplineCurve3D(BSplineCurve):
         self._bbox = new_bounding_box
 
     def _bounding_box(self):
-        bbox = self.curve.bbox
-        return volmdlr.core.BoundingBox(bbox[0][0], bbox[1][0],
-                                        bbox[0][1], bbox[1][1],
-                                        bbox[0][2], bbox[1][2])
+        """Creates a bounding box from the bspline points."""
+        return volmdlr.core.BoundingBox.from_points(self.discretization_points())
 
     def look_up_table(self, resolution: int = 20, start_parameter: float = 0,
                       end_parameter: float = 1):
         """
-        Creates a table of equivalence between the parameter t (eval. of the BSplineCurve) and the cumulative distance.
+        Creates a table of equivalence between parameter t (evaluation of BSplineCurve) and the cumulative distance.
 
         :param resolution: The precision of the table. Auto-adjusted by the
             algorithm. Default value set to 20
@@ -4390,6 +4400,10 @@ class BSplineCurve3D(BSplineCurve):
         return normal
 
     def get_direction_vector(self, abscissa=0.0):
+        """
+        Calculates direction vector at given abscissa value (value between o and bspline length).
+
+        """
         length = self.length()
         if abscissa >= length:
             abscissa2 = length
@@ -4403,6 +4417,10 @@ class BSplineCurve3D(BSplineCurve):
         return tangent
 
     def direction_vector(self, abscissa=0.):
+        """
+        Gets direction vector at given abscissa value (value between o and bspline length).
+
+        """
         if not self._direction_vector_memo:
             self._direction_vector_memo = {}
         if abscissa not in self._direction_vector_memo:
@@ -4573,6 +4591,12 @@ class BSplineCurve3D(BSplineCurve):
         return bspline_curve.__class__.from_points_interpolation(local_discretization, bspline_curve.degree)
 
     def trim_between_evaluations(self, parameter1: float, parameter2: float):
+        """
+        Trims the Bspline between two abscissa evaluation parameters.
+
+        :param parameter1: evaluation parameter 1, bigger than 0 and smaller than its length.
+        :param parameter2: evaluation parameter 2, bigger than 0 and smaller than its length.
+        """
         warnings.warn('Use BSplineCurve3D.trim instead of trim_between_evaluation')
         parameter1, parameter2 = min([parameter1, parameter2]), \
             max([parameter1, parameter2])
@@ -5165,7 +5189,7 @@ class Arc3D(ArcMixin, Edge):
         u3.normalize()
         u4 = other_arc.circle.normal.cross(u3)
 
-        r1, r2 = self.circle.radius, other_arc.circle.radius
+        radius1, radius2 = self.circle.radius, other_arc.circle.radius
 
         a, b, c, d = u1.dot(u1), u1.dot(u2), u1.dot(u3), u1.dot(u4)
         e, f, g = u2.dot(u2), u2.dot(u3), u2.dot(u4)
@@ -5174,27 +5198,27 @@ class Arc3D(ArcMixin, Edge):
         k, l, m, n, o = w.dot(u1), w.dot(u2), w.dot(u3), w.dot(u4), w.dot(w)
 
         def distance_squared(x):
-            return (a * ((math.cos(x[0])) ** 2) * r1 ** 2 + e * (
-                    (math.sin(x[0])) ** 2) * r1 ** 2
-                    + o + h * ((math.cos(x[1])) ** 2) * r2 ** 2 + j * (
-                            (math.sin(x[1])) ** 2) * r2 ** 2
-                    + b * math.sin(2 * x[0]) * r1 ** 2 - 2 * r1 * math.cos(
+            return (a * ((math.cos(x[0])) ** 2) * radius1 ** 2 + e * (
+                    (math.sin(x[0])) ** 2) * radius1 ** 2
+                    + o + h * ((math.cos(x[1])) ** 2) * radius2 ** 2 + j * (
+                            (math.sin(x[1])) ** 2) * radius2 ** 2
+                    + b * math.sin(2 * x[0]) * radius1 ** 2 - 2 * radius1 * math.cos(
                         x[0]) * k
-                    - 2 * r1 * r2 * math.cos(x[0]) * math.cos(x[1]) * c
-                    - 2 * r1 * r2 * math.cos(x[0]) * math.sin(
-                        x[1]) * d - 2 * r1 * math.sin(x[0]) * l
-                    - 2 * r1 * r2 * math.sin(x[0]) * math.cos(x[1]) * f
-                    - 2 * r1 * r2 * math.sin(x[0]) * math.sin(
-                        x[1]) * g + 2 * r2 * math.cos(x[1]) * m
-                    + 2 * r2 * math.sin(x[1]) * n + i * math.sin(
-                        2 * x[1]) * r2 ** 2)
+                    - 2 * radius1 * radius2 * math.cos(x[0]) * math.cos(x[1]) * c
+                    - 2 * radius1 * radius2 * math.cos(x[0]) * math.sin(
+                        x[1]) * d - 2 * radius1 * math.sin(x[0]) * l
+                    - 2 * radius1 * radius2 * math.sin(x[0]) * math.cos(x[1]) * f
+                    - 2 * radius1 * radius2 * math.sin(x[0]) * math.sin(
+                        x[1]) * g + 2 * radius2 * math.cos(x[1]) * m
+                    + 2 * radius2 * math.sin(x[1]) * n + i * math.sin(
+                        2 * x[1]) * radius2 ** 2)
 
         x01 = npy.array([self.angle / 2, other_arc.angle / 2])
 
         res1 = least_squares(distance_squared, x01, bounds=[(0, 0), (self.angle, other_arc.angle)])
 
-        point1 = self.point_at_abscissa(res1.x[0] * r1)
-        point2 = other_arc.point_at_abscissa(res1.x[1] * r2)
+        point1 = self.point_at_abscissa(res1.x[0] * radius1)
+        point2 = other_arc.point_at_abscissa(res1.x[1] * radius2)
 
         return point1, point2
 
@@ -5408,8 +5432,13 @@ class Arc3D(ArcMixin, Edge):
         :return:
         """
         new_faces = []
-        _, section_contour_3d = args
-        for contour_primitive in section_contour_3d.primitives:
+        section_contour2d, section_contour3d = args
+        if section_contour3d is None:
+            start_tangent = self.unit_direction_vector(0.)
+            normal = self.unit_normal_vector(0.)
+            tangent_normal_orthonormal = start_tangent.cross(normal)
+            section_contour3d = section_contour2d.to_3d(self.start, normal, tangent_normal_orthonormal)
+        for contour_primitive in section_contour3d.primitives:
             new_faces.extend(contour_primitive.revolution(
                 self.circle.center, self.circle.normal, self.angle))
         return new_faces
@@ -5425,8 +5454,8 @@ class FullArc3D(FullArcMixin, Arc3D):
                  name: str = ''):
         self._utd_frame = None
         self._bbox = None
-        FullArcMixin.__init__(self, circle=circle, start_end=start_end, name=name)
-        Arc3D.__init__(self, circle=circle, start=start_end, end=start_end)
+        FullArcMixin.__init__(self, circle=circle, start_end=start_end)
+        Arc3D.__init__(self, circle=circle, start=start_end, end=start_end, name=name)
 
     def __hash__(self):
         return hash(('Fullarc3D', self.circle, self.start_end))
@@ -5527,8 +5556,8 @@ class FullArc3D(FullArcMixin, Arc3D):
         :return: A new FullArc3D object that is the result of the translation.
         :rtype: FullArc3D.
         """
-        new_start_end = self.start.translation(offset, True)
-        new_circle = self.circle.translation(offset, True)
+        new_start_end = self.start.translation(offset)
+        new_circle = self.circle.translation(offset)
         return FullArc3D(new_circle, new_start_end, name=self.name)
 
     def frame_mapping(self, frame: volmdlr.Frame3D, side: str):
@@ -5888,6 +5917,7 @@ class ArcEllipse3D(Edge):
                 self.__class__(self.ellipse, split_point, self.end)]
 
     def get_reverse(self):
+        """Gets the same ellipse but in the reverse direction."""
         new_frame = volmdlr.Frame3D(self.ellipse.frame.origin, self.ellipse.frame.u, -self.ellipse.frame.v,
                                     self.ellipse.frame.u.cross(-self.ellipse.frame.v))
         ellipse3d = volmdlr_curves.Ellipse3D(self.ellipse.major_axis, self.ellipse.minor_axis, new_frame)
@@ -6000,7 +6030,7 @@ class FullArcEllipse3D(FullArcEllipse, ArcEllipse3D):
         :return: list of two Arc of ellipse.
         """
         if split_point.is_close(self.start, tol) or split_point.is_close(self.end, tol):
-            raise ValueError("Point should be different of start and end.")
+            return [self, None]
         if not self.point_belongs(split_point, 1e-5):
             raise ValueError("Point not on the ellipse.")
         return [ArcEllipse3D(self.ellipse, self.start_end, split_point),