+++ volmdlr/faces.py
@@ -44,7 +44,8 @@ def octree_face_decomposition(face):
     Decomposes the face discretization triangle faces inside eight boxes from a bounding box octree structure.
 
     :param face: given face.
-    :return: returns a dictionary containing bounding boxes as keys and as values, a list of faces inside that bbox.
+    :return: returns a dictionary containing bounding boxes as keys and as values, a list of faces
+    inside that bounding box.
     """
     triangulation = face.triangulation()
     triangulation_faces = triangulation.faces
@@ -1939,9 +1940,9 @@ class CylindricalFace3D(Face3D):
         :param arc: Arc3D to be verified.
         :return: True if it is inside, False otherwise.
         """
-        if not math.isclose(abs(arc.frame.w.dot(self.surface3d.frame.w)), 1.0, abs_tol=1e-6):
+        if not math.isclose(abs(arc.circle.frame.w.dot(self.surface3d.frame.w)), 1.0, abs_tol=1e-6):
             return False
-        if not math.isclose(self.radius, arc.radius, abs_tol=1e-6):
+        if not math.isclose(self.radius, arc.circle.radius, abs_tol=1e-6):
             return False
         return self.arcellipse_inside(arc)