+++ volmdlr/wires.py
@@ -189,6 +189,10 @@ class WireMixin:
     def _data_hash(self):
         return sum(hash(e) for e in self.primitives) + len(self.primitives)
 
+    def to_dict(self, *args, **kwargs):
+        """Avoids storing points in memo that makes serialization slow."""
+        return PhysicalObject.to_dict(self, use_pointers=False)
+
     def length(self):
         if not self._length:
             length = 0.
@@ -536,7 +540,7 @@ class WireMixin:
 
             while to_continue:
                 broke = False
-                for p, primitive in enumerate(list_edges):
+                for index_primitive, primitive in enumerate(list_edges):
 
                     if primitive.is_point_edge_extremity(new_primitives[i][-1].end, tol):
                         if new_primitives[i][-1].end.is_close(primitive.start, tol):
@@ -556,7 +560,7 @@ class WireMixin:
                         broke = True
                         break
 
-                if ((not broke) and (len(list_edges) == p + 1)) or len(list_edges) == 0:
+                if ((not broke) and (len(list_edges) == index_primitive + 1)) or len(list_edges) == 0:
                     to_continue = False
 
         wires = [cls(primitives_wire) for primitives_wire in new_primitives]
@@ -630,6 +634,33 @@ class WireMixin:
             if not self.primitive_over_wire(primitive_2):
                 return False
         return True
+
+    @classmethod
+    def from_circle(cls, circle):
+        """
+        Creates a Contour from a circle.
+
+        :param circle: Circle
+        :return:
+        """
+        point = circle.point_at_abscissa(0.0)
+        return cls([circle.trim(point, point)])
+
+    def plot(self, ax=None, edge_style=EdgeStyle()):
+        """Wire 2D plot using Matplotlib."""
+        if ax is None:
+            ax = self._get_plot_ax()
+
+        if edge_style.equal_aspect:
+            ax.set_aspect('equal')
+
+        for element in self.primitives:
+            element.plot(ax=ax, edge_style=edge_style)
+
+        ax.margins(0.1)
+        plt.show()
+
+        return ax
 
 
 class EdgeCollection3D(WireMixin):
@@ -725,10 +756,6 @@ class Wire2D(WireMixin, PhysicalObject):
 
     def __hash__(self):
         return hash(('wire2d', tuple(self.primitives)))
-
-    def to_dict(self, *args, **kwargs):
-        """Avoids storing points in memo that makes serialization slow."""
-        return PhysicalObject.to_dict(self, use_pointers=False)
 
     def area(self):
         """ Gets the area for a Wire2D."""
@@ -1257,34 +1284,25 @@ class Wire2D(WireMixin, PhysicalObject):
         return self.__class__([primitive.frame_mapping(frame, side)
                                for primitive in self.primitives])
 
-    def plot(self, ax=None, edge_style=EdgeStyle()):
-        """Wire 2D plot using Matplotlib."""
-        if ax is None:
-            _, ax = plt.subplots()
-
-        if edge_style.equal_aspect:
-            ax.set_aspect('equal')
-
-        for element in self.primitives:
-            element.plot(ax=ax, edge_style=edge_style)
-
-        ax.margins(0.1)
-        plt.show()
-
+    def _get_plot_ax(self):
+        _, ax = plt.subplots()
         return ax
 
 
-class Wire3D(WireMixin, volmdlr.core.CompositePrimitive3D):
+class Wire3D(WireMixin, PhysicalObject):
     """
     A collection of simple primitives, following each other making a wire.
 
     """
 
-    def __init__(self, primitives: List[volmdlr.core.Primitive3D],
+    def __init__(self, primitives: List[volmdlr.core.Primitive3D], color=None, alpha: float = 1.0,
                  name: str = ''):
         self._bbox = None
         self._length = None
-        volmdlr.core.CompositePrimitive3D.__init__(self, primitives=primitives, name=name)
+        self.primitives = primitives
+        self.color = color
+        self.alpha = alpha
+        PhysicalObject.__init__(self, name=name)
 
     def _bounding_box(self):
         """
@@ -1418,8 +1436,40 @@ class Wire3D(WireMixin, volmdlr.core.CompositePrimitive3D):
                      in self.primitives]
         return Wire3D(new_edges, self.name)
 
+    def _get_plot_ax(self):
+        fig = plt.figure()
+        ax = fig.add_subplot(111, projection='3d')
+        return ax
 
-# TODO: define an edge as an opened polygon and allow to compute area from this reference
+    def babylon_points(self):
+        """
+        Returns a list of discretization points from the 3D primitive.
+        """
+        points = []
+        if hasattr(self, 'primitives') and hasattr(self.primitives[0], "discretization_points"):
+            for primitive in self.primitives:
+                points.extend([*point] for point in primitive.discretization_points())
+        elif hasattr(self, "discretization_points"):
+            points.extend([*point] for point in self.discretization_points())
+        return points
+
+    def babylon_lines(self, points=None):
+        if points is None:
+            points = self.babylon_points()
+        babylon_lines = {'points': points,
+                         'alpha': self.alpha,
+                         'name': self.name,
+                         'color': list(self.color) if self.color is not None else [0.8, 0.8, 0.8]
+                         }
+        return [babylon_lines]
+
+    def babylon_curves(self):
+        points = self.babylon_points()
+        if points:
+            babylon_curves = self.babylon_lines(points)[0]
+            return babylon_curves
+        return None
+
 
 class ContourMixin(WireMixin):
     """
@@ -1697,7 +1747,7 @@ class ContourMixin(WireMixin):
         """
         Check if two contour are adjacent.
 
-        So: are sharing primitives but not supperposing or none is inside the other.
+        So: are sharing primitives but not superposing or none is inside the other.
         """
 
         if (self.is_inside(contour) or contour.is_inside(self)
@@ -1844,7 +1894,7 @@ class ContourMixin(WireMixin):
         Return extremities points of a list of points on a contour.
 
         """
-        # TODO: rewrite this awfull code!
+        # TODO: rewrite this awful code!
         points = []
         primitives = self.primitives
         for prim in primitives:
@@ -1883,7 +1933,7 @@ class ContourMixin(WireMixin):
 
     def primitive_section_over_contour(self, primitive, abs_tol: float = 1e-6):
         """
-        Verifies if at least a small section of a primitive is over a contour, not necessarilly the entire primitive.
+        Verifies if at least a small section of a primitive is over a contour, not necessarily the entire primitive.
 
         """
         for prim in self.primitives:
@@ -1981,7 +2031,7 @@ class ContourMixin(WireMixin):
         """
         Create a new contour from self, but starting at given point.
 
-        :param point: othe point.
+        :param point: other point.
         :return: new contour
         """
         new_primitives_order = []
@@ -2374,13 +2424,12 @@ class Contour2D(ContourMixin, Wire2D):
                                      volmdlr.Point2D(xi, 1))
 
             iteration_contours2 = []
-            for c in iteration_contours:
-                sc = c.cut_by_line(cut_line)
-                lsc = len(sc)
-                if lsc == 1:
-                    cutted_contours.append(c)
+            for contour in iteration_contours:
+                split_contours = contour.cut_by_line(cut_line)
+                if len(split_contours) == 1:
+                    cutted_contours.append(contour)
                 else:
-                    iteration_contours2.extend(sc)
+                    iteration_contours2.extend(split_contours)
 
             iteration_contours = iteration_contours2[:]
         cutted_contours.extend(iteration_contours)
@@ -2539,7 +2588,7 @@ class Contour2D(ContourMixin, Wire2D):
 
     def discretized_contour(self, n: float):
         """
-        Discretize each contour's primitive and return a new contour with teses discretized primitives.
+        Discretize each contour's primitive and return a new contour with theses discretized primitives.
         """
         contour = Contour2D((self.discretized_primitives(n)))
 
@@ -2866,12 +2915,12 @@ class ClosedPolygon2D(ClosedPolygonMixin, Contour2D):
                          - sum(i * j for i, j in zip(y, x1)))
 
     def center_of_mass(self):
-        lp = len(self.points)
-        if lp == 0:
+        lngth_points = len(self.points)
+        if lngth_points == 0:
             return volmdlr.O2D
-        if lp == 1:
+        if lngth_points == 1:
             return self.points[0]
-        if lp == 2:
+        if lngth_points == 2:
             return 0.5 * (self.points[0] + self.points[1])
 
         x = [point.x for point in self.points]
@@ -3462,8 +3511,8 @@ class ClosedPolygon2D(ClosedPolygonMixin, Contour2D):
                 point.plot(ax=ax, color=edge_style.color, alpha=edge_style.alpha)
 
         if point_numbering:
-            for ip, point in enumerate(self.points):
-                ax.text(*point, f'point {ip + 1}', ha='center', va='top')
+            for index_point, point in enumerate(self.points):
+                ax.text(*point, f'point {index_point + 1}', ha='center', va='top')
 
         if edge_style.equal_aspect:
             ax.set_aspect('equal')
@@ -3517,11 +3566,8 @@ class ClosedPolygon2D(ClosedPolygonMixin, Contour2D):
         """
         x_min, x_max, y_min, y_max = self.bounding_rectangle.bounds()
 
-        n = number_points_x + 2
-        m = number_points_y + 2
-
-        x = npy.linspace(x_min, x_max, num=n)
-        y = npy.linspace(y_min, y_max, num=m)
+        x = npy.linspace(x_min, x_max, num=number_points_x + 2)
+        y = npy.linspace(y_min, y_max, num=number_points_y + 2)
 
         grid_point_index = {}